[joblib] split _core module out
